DjangoApi/appleMap/requirements.txt
absl-py==0.8.1
appdirs==1.4.0
asgiref==3.2.3
astor==0.8.0
backports.weakref==1.0rc1
bleach==1.5.0
cachetools==3.1.1
certifi==2019.11.28
chardet==3.0.4
cycler==0.10.0
Django==3.0
django-cors-headers==3.2.0
django-crispy-forms==1.8.1
djangorestframework==3.11.0
gast==0.2.2
google-auth==1.7.2
google-auth-oauthlib==0.4.1
google-pasta==0.1.8
grpcio==1.25.0
h5py==2.10.0
html5lib==0.9999999
idna==2.8
joblib==0.14.0
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
Markdown==3.1.1
matplotlib==3.1.1
numpy==1.17.2
oauthlib==3.1.0
opt-einsum==3.1.0
packaging==16.8
pandas==0.25.2
Pillow==6.2.1
protobuf==3.11.1
pyasn1==0.4.8
pyasn1-modules==0.2.7
pyparsing==2.1.10
python-dateutil==2.8.0
pytz==2019.2
PyYAML==5.2
requests==2.22.0
requests-oauthlib==1.3.0
rsa==4.0
scikit-learn==0.21.3
scipy==1.1.0
selenium==3.141.0
six==1.10.0
sqlparse==0.3.0
tensorboard==2.0.2
tensorflow==2.0.0
tensorflow-estimator==2.0.1
termcolor==1.1.0
urllib3==1.25.7
virtualenv==16.7.5
Werkzeug==0.16.0
wrapt==1.11.2
